CIF-961 - Shopping cart (MiniCart) component using GraphQL API
Get rid of the initial container component, it brought nothing to the table.
