+++ ui.apps/src/main/javascript/minicart/src/components/Minicart/index.js
@@ -11,4 +11,4 @@
  *    governing permissions and limitations under the License.
  *
  ******************************************************************************/
-export { default } from './container';
+export { default } from './minicart';
